pyproject.toml
[tool.poetry]
name = "test-ctls"
version = "0.1.0"
description = ""
authors = ["Alex_Goryachev <alexey.i.goryachev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
